Extract open neighbors as a memoized selector
